refactor: lift platform modules out of core ahead of the n23 rename
Preparation for #2093, which renames gyrinx.core/gyrinx.content to
n23.core/n23.content. Three modules under gyrinx.core are not
edition-specific and would be renamed into the edition namespace by that
move, so they go to the platform package first.

- gyrinx.core.badges -> gyrinx.badges (Patreon tier badges; gyrinx.api
  already depended on it)
- TinyMCE widget and config -> gyrinx.widgets (gyrinx.pages already
  depended on it); ColorRadioSelect stays in core, since it renders
  core-owned templates
- gyrinx.core.models.state_machine -> gyrinx.state_machine, so the task
  framework no longer imports the edition app for a generic utility
- tasks/checks.py treats n23.*/n26.* as first-party, so the rename can't
  silently narrow @task auto-discovery and weaken the #1947 guard

No schema change: the generated transition models take __module__ and
app_label from the owning model, not from the state_machine module, so
makemigrations --check stays clean.
