+++ gyrinx/api/patreon.py
@@ -2,7 +2,7 @@ import logging
 
 from allauth.account.models import EmailAddress
 
-from gyrinx.core.badges import rank_for_tier_title
+from gyrinx.badges import rank_for_tier_title
 from gyrinx.core.models.auth import PatreonStatus, UserProfile
 
 logger = logging.getLogger(__name__)
