+++ gyrinx/core/forms/pack.py
@@ -23,7 +23,7 @@ from gyrinx.core.models.pack import (
     CustomContentPack,
     CustomContentPackAttachment,
 )
-from gyrinx.core.widgets import TINYMCE_EXTRA_ATTRS, TinyMCEWithUpload
+from gyrinx.widgets import TINYMCE_EXTRA_ATTRS, TinyMCEWithUpload
 from gyrinx.forms import group_select
 from gyrinx.models import FighterCategoryChoices, equipment_category_groups
