+++ gyrinx/core/views/battle.py
@@ -25,7 +25,7 @@ from gyrinx.core.handlers.crew import crew_spread_rating, crew_stash_totals
 from gyrinx.core.models import Battle, Campaign, CampaignAction
 from gyrinx.core.models.crew import Crew
 from gyrinx.core.models.events import EventNoun, EventVerb, log_event
-from gyrinx.core.models.state_machine import InvalidStateTransition
+from gyrinx.state_machine import InvalidStateTransition
 from gyrinx.core.utils import get_return_url, safe_redirect
